systemtest/requirements.txt
assertpy==1.1
celery==4.4.7
mysqlclient==2.0.1
pymongo==3.10.1
pytest==6.1.0
python-engineio==4.0.0
python-socketio==5.0.2
redis==3.4.1
requests==2.24.0
socketIO-client==0.7.2